requirements.txt
# Required
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
aiohttp>=3.9.0
asyncio

# LLM & Research
openai>=1.6.0
asknews>=0.3.0

# Data processing
numpy>=1.26.0

# Optional - for advanced search
# tavily-python>=0.5.0
# exa-py>=1.0.0
